Revert "t_point<double>"
This reverts commit 0b1413cbe8f7120580ce476b2b1677dd3d681e8b.

+++ src/types.h
@@ -3,5 +3,5 @@
 #include <boost/geometry/geometries/polygon.hpp>
 #include <boost/geometry/geometries/point_xy.hpp>
 
-typedef boost::geometry::model::d2::point_xy<double> t_point;
+typedef boost::geometry::model::d2::point_xy<int> t_point;
 typedef boost::geometry::model::polygon<t_point> t_polygon;
